// src/api/openapi/models/ErrorCode.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API Merin
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ErrorCode = {
    UNKNOWN: 'UNKNOWN',
    BANNED: 'BANNED',
    UNSUPPORTED_ROLE: 'UNSUPPORTED_ROLE',
    UNKNOWN_ROLE: 'UNKNOWN_ROLE',
    LOGIN_FAILED: 'LOGIN_FAILED',
    USER_EXISTS: 'USER_EXISTS',
    UNKNOWN_EMAIL: 'UNKNOWN_EMAIL',
    UNKNOWN_USER: 'UNKNOWN_USER',
    WRONG_PASSWORD: 'WRONG_PASSWORD',
    INVALID_PASSWORD_TOKEN: 'INVALID_PASSWORD_TOKEN',
    INACTIVE: 'INACTIVE'
} as const;
export type ErrorCode = typeof ErrorCode[keyof typeof ErrorCode];


export function instanceOfErrorCode(value: any): boolean {
    for (const key in ErrorCode) {
        if (Object.prototype.hasOwnProperty.call(ErrorCode, key)) {
            if (ErrorCode[key as keyof typeof ErrorCode] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ErrorCodeFromJSON(json: any): ErrorCode {
    return ErrorCodeFromJSONTyped(json, false);
}

export function ErrorCodeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ErrorCode {
    return json as ErrorCode;
}

export function ErrorCodeToJSON(value?: ErrorCode | null): any {
    return value as any;
}

export function ErrorCodeToJSONTyped(value: any, ignoreDiscriminator: boolean): ErrorCode {
    return value as ErrorCode;
}